+++ src/io/mod.rs
@@ -12,6 +12,7 @@ pub mod guidance;
 pub mod guidance_engine;
 pub mod input;
 pub mod output;
+pub mod parse;
 pub mod schema;
 #[cfg(test)]
 mod test;